feat(server): #5 declare loss if user disconnects mid game
Previously, if a user disconnected from the websocket connection they were
simply removed from the lobby. If this happened during a progressing
game, the user that was left alone in the lobby was left hanging forever.
Now the user that is left in the lobby is declared the winner of the
game and the leaver is declared the loser.