feat(server): #11 use trivia api key to guarantee unique questions
Previously there was no guarantee that the questions that were requested
from the Trivia API were going to be unique (between different
requests).

Now a trivia API key is also requested at the start of a game session
that is then used to request all other questions. This API key
guarantees that there will be no duplicate questions returned during the
lifetime of the game session.