feat(server): #15 implement maximum game length
Now all games have a maximum length configured in the settings, which is
sent to the client. The server now expects to receive a "game.timeout"
event whenever the maximum game length expires, server allows users to
finish the last question they were on and then ends the game.

Also, it is no longer "game.start" event's responsibility to notify the
client to continue to the next question, instead the client should
expect to receive a "question.next" event separately.