feat(client): #15 add a header for the game timer, player names and healths
Implemented a UI component to display the game timer, as well as
player's and opponent's name and health. When the timer expires a
"game.timeout" event is sent to notify the server that maximum game time
has expired.

Also added some CSS to center screen contents to make it look a bit
nicer.