refactor(server): #17 refactor GameConsumer authentication
Previously a random string token was being generated on lobby creation
and joins, this token was stored in with the lobby and used to
authenticate users connecting to GameConsumer. But using this method
created a problem where if a second user decided not to or failed to
connect to the GameConsumer after "joining" using the API, the lobby
would be left "hanging" and no other users would be able to join.
Because there was no mechanism to remove an non-connected user's token
from the lobby.

Now instead of a string token, a very short lived (5s) jwt token is
generated, no token is stored with the lobby any longer, the users
simply provide the token when connecting to the GameConsumer and the
server can validate user identities by verifying the jwt signature. This
approach circumvents the hanging lobby problem.