feat(server): validate user by lobby_name claim on GameConsumer connect
Added a lobby_name claim to the lobby authentication token. This claim
is used to validate that the user is authenticated to connect to the
lobby that they are trying to connect to.

This prevents users from using lobby authentication tokens to join any
lobby they want. Now, the token is generated for a specific lobby and
can be only be used to join that specific lobby.